requirements.txt
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
cssselect==1.1.0
Django==3.2.4
django-jsonfield==1.4.1
djangorestframework==3.12.4
fake-useragent==0.1.11
idna==2.10
lxml==4.6.3
parse==1.19.0
pyee==8.1.0
pyppeteer==0.2.5
pyquery==1.4.3
pytz==2021.1
requests==2.25.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tqdm==4.61.1
urllib3==1.26.5
w3lib==1.22.0
websockets==8.1
